+++ Native/Win32.cs
@@ -148,6 +148,17 @@ public static class Win32
         [MarshalAs(UnmanagedType.LPArray, SizeParamIndex = 1)] RAWINPUTDEVICE[] pRawInputDevices,
         uint uiNumDevices, uint cbSize);
 
+    [StructLayout(LayoutKind.Sequential)]
+    public struct RAWINPUTDEVICELIST
+    {
+        public IntPtr hDevice;
+        public uint dwType;
+    }
+
+    [DllImport("user32.dll", SetLastError = true)]
+    public static extern uint GetRawInputDeviceList(
+        [In, Out] RAWINPUTDEVICELIST[]? pRawInputDeviceList, ref uint puiNumDevices, uint cbSize);
+
     [DllImport("user32.dll")]
     public static extern uint GetRawInputData(IntPtr hRawInput, uint uiCommand,
         IntPtr pData, ref uint pcbSize, uint cbSizeHeader);